fix for later-than-in-next-HaskellPlatform stupid GL newtypes it involves disputable API-change-or-is-that-just-the- -package-splitting, and, realToFrac has slowness vs. correctness problems and we don't have proper conversion operators yet! er. this is a bit of an awful patch.

+++ Main.hs
@@ -140,5 +140,5 @@ positionToZeroOneRange :: Position -> IO (Double,Double)
 positionToZeroOneRange (Position x y) = get viewport >>=
 	\(Position minX minY, Size sizeX sizeY) -> return
 		(fromIntegral (x - minX) / fromIntegral sizeX,
-		 fromIntegral (sizeY - (y - minY) {-stupid GLUT difference of *origin* to OpenGL-}) / fromIntegral sizeY)
+		 fromIntegral (fromIntegral sizeY - (y - minY) {-stupid GLUT difference of *origin* to OpenGL-}) / fromIntegral sizeY)
